makefile##
This program is about to marge the System program with kernel compiling beheavior at the very first time of the System got Installed a Linux Distro on a System. Its a part of "Independent System arch Linux Distro 32-bit". This progream is aparently succcessful based on executed-on first test on a Live System(32-Bit)