rmb32_Refresh_assist
This program is a Python program to providing the REFRESH option like MS-Windows OS in a Linux Distributiion. But the main problem is Its need to assign into a system after installation as a live system. I would like to attach it within an Linux Os under the path /home/${username/nautils-script/scripts/ . But can't find any helpful way to do. The program is successfully executed under Python-2*, 3* env. May please kindly help me in this matter. Thanking you in advance.